+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This file is used to list changes made in each version of the openstack-block-st
 * Update deprecated keystoneclient.middleware.auth_token in api-paste.ini
 * Add attributes for quota_snapshots, no_snapshot_gb_quota and use_default_quota_class
 * Add glance_api_insecure and make glance_api_servers to be prefixed with scheme
+* Add sysfsutils package for storwize/SVC driver
 
 ## 10.0.0
 * Upgrading to Juno